Initial version for the URL paired task
1. Get URLs one line at a time from an input file
2. 'sanitise' them, meaning remove the only substring from them that is
common in the pairs
3. Attempt to load each and every such 'sanitised' URL into a HashSet.
4. If the add() fails, it means we found a pair.
4.1 Put the sanitised version of this URL pair into a LinkedList<String>
duplicatesURLs
4.2 Remove this sanitised version of URL from the HastSet
5. When this is finished we have
5.1 All Single URLs in the HashSet and all Paired URLs in the LinkedList

6. I just sysout them, regardless of order and the original data, but
from this point on it is just a matter of formatting to restore the
original state of the URLs, and restore order in which they appeared in
the file (if that is even necessary)

**********************************
NOTE: this solution of course supposed that the specification was
correct and complete in regards to the input data, meaning that the
pairs are recognisable ALWAYS by that particular part of the URL (_rq
and _rp) and that these are ALWAYS right before the .xml extension part.
If the specification was misunderstood this might not be the solution we
were looking for.